# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Gastronómico",
    "version": "1.0.0",
    "description": "Documentación de la API Multi-Tenant para Restaurantes",
    "contact": {
      "name": "Soporte API",
      "email": "api@gastronomico.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/v1",
      "description": "Servidor local"
    },
    {
      "url": "https://backend-gastronomico.onrender.com/v1",
      "description": "Servidor desplegado"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtenido del endpoint /auth/login"
      }
    }
  },
...
  "paths": {
...
    "/plan": {
      "post": {
        "summary": "Crear un nuevo plan de suscripción",
        "description": "Crea un nuevo plan con características y precios asociados",
        "tags": [
          "Plan"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "billingCycleId"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Plan Pro"
                  },
                  "description": {
                    "type": "string",
                    "example": "Plan profesional con todas las características"
                  },
                  "isFree": {
                    "type": "boolean",
                    "example": false
                  },
                  "billingCycleId": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "isFree": {
                      "type": "boolean"
                    },
                    "billingCycleId": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "get": {
        "summary": "Obtener todos los planes",
        "description": "Lista todos los planes de suscripción disponibles",
        "tags": [
          "Plan"
        ],
        "responses": {
          "200": {
            "description": "Lista de planes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer"
                    },
                    "rows": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/plan/{id}": {
      "get": {
        "summary": "Obtener un plan por ID",
        "tags": [
          "Plan"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan encontrado"
          },
          "404": {
            "description": "Plan no encontrado"
          }
        }
      },
      "patch": {
        "summary": "Actualizar un plan",
        "tags": [
          "Plan"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "isFree": {
                    "type": "boolean"
                  },
                  "billingCycleId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan actualizado"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Plan no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un plan",
        "tags": [
          "Plan"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan eliminado"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Plan no encontrado"
          }
        }
      }
    },
    "/plan/{id}/status": {
      "patch": {
        "summary": "Actualizar estado del plan",
        "tags": [
          "Plan"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "statusId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado"
          },
          "400": {
            "description": "statusId es requerido"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Plan no encontrado"
          }
        }
      }
    },
...
    "/plan-features/plan/{planId}": {
      "get": {
        "summary": "Obtener características de un plan",
        "description": "Lista todas las características de un plan específico",
        "tags": [
          "PlanFeatures"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "planId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de características",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer"
                    },
                    "rows": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/plan-price/plan/{planId}": {
      "get": {
        "summary": "Obtener precios de un plan específico",
        "description": "Lista todos los precios en diferentes monedas de un plan",
        "tags": [
          "PlanPrice"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "planId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de precios del plan"
          },
          "404": {
            "description": "Plan no encontrado"
          }
        }
      }
    },
...
  }
}